% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binnedTable}
\alias{binnedTable}
\title{binnedTable}
\format{
A matrix
}
\usage{
binnedTable
}
\description{
A matrix of segment mean copy number values per bin or probe x sample
}
\keyword{datasets}
