% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber.R, R/readFunctions.R
\name{calculateRelatednessCn}
\alias{calculateRelatednessCn}
\title{Calculate relatedness scores for paired tumours}
\usage{
calculateRelatednessCn(
  segmentTable,
  pairs,
  reference = NULL,
  cnType = c("alleleSpecific", "VCF"),
  excludeChromosomes = "Y",
  maxgap = NULL
)

calculateRelatednessCn(
  segmentTable,
  pairs,
  reference = NULL,
  cnType = c("alleleSpecific", "VCF"),
  excludeChromosomes = "Y",
  maxgap = NULL
)
}
\arguments{
\item{segmentTable}{A segment table generated by the readAlleleSpecific or readVCFCn functions.}

\item{pairs}{A table of paired samples from the dataset, to test for relatedness.}

\item{reference}{A numeric vector of pair scores comprising the reference distribution, generated from the \code{makeReferenceCN} function. If omitted, p-value calculation will be skipped.}

\item{cnType}{The type of copy number data provided. Currently supported options are a custom allele specific data format, and standard copy number VCF files.}

\item{excludeChromosomes}{The name(s) of any chromosomes to be excluded.}

\item{maxgap}{The maximum gap between two breakpoints for them to be considered concordant. If unspecified, it is automatically set to 5 times the average interprobe distance of the assay.}
}
\value{
A data frame listing the tumour pairs contained in \code{pairs}, their relatedness scores and p-values for relatedness.

A data frame listing the tumour pairs contained in \code{pairs}, their relatedness scores and p-values for relatedness.
}
\description{
Calculates the relatedness scores and (optionally) p-values for paired tumours from copy number data

Calculates the relatedness scores and (optionally) p-values for paired tumours from copy number data
}
