% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{calculateRelatednessMutations}
\alias{calculateRelatednessMutations}
\title{Calculate relatedness scores for paired tumours}
\usage{
calculateRelatednessMutations(
  mutationTable,
  pairs,
  additionalMutations = NULL,
  nAdditionalSamples = 0,
  reference = NULL,
  excludeChromosomes = "chrY",
  scaleAFs = FALSE
)
}
\arguments{
\item{mutationTable}{A table of mutations in each sample and their allele frequencies.}

\item{pairs}{A table of paired samples from the dataset, to test for relatedness.}

\item{additionalMutations}{A table of mutations to be taken into account when calculating population frequencies. At a minimum, a table of the mutations in the population being studied. This is more informative when tumour type-specific mutations are included from external sources (e.g. TCGA).}

\item{nAdditionalSamples}{The number of samples used to derive the additional mutations table.}

\item{reference}{A numeric vector of pair scores comprising the reference distribution, generated from the \code{makeReferenceMutations} function. If omitted, p-value calculation will be skipped.}

\item{excludeChromosomes}{The name(s) of any chromosomes to be excluded.}

\item{scaleAFs}{Scale AFs per-sample by the highest AF within each sample. Only recommended for data with significant normal contamination that you are confident contains at least one clonal mutation per sample.}
}
\value{
A data frame listing the tumour pairs contained in \code{pairs}, their relatedness scores and p-values for relatedness.
}
\description{
Calculates the relatedness scores and (optionally) p-values for paired tumours from mutation data
}
