% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber.R
\name{getSharedBreaks}
\alias{getSharedBreaks}
\title{Generate the shared breakpoints per pair calculated by breakclone and used to assess relatedness.}
\usage{
getSharedBreaks(
  segmentTable,
  pairs,
  outdir = ".",
  cnType = c("alleleSpecific", "VCF"),
  maxgap = NULL,
  save = FALSE
)
}
\arguments{
\item{segmentTable}{A segment table generated by the readAlleleSpecific or readVCFCn functions.}

\item{pairs}{A table of paired samples from the dataset. All tumours present in this table will be paired with all tumours from other patients.}

\item{outdir}{A path to save the shared breakpoints. If unspecified, it is automatically set to the working directory.}

\item{cnType}{The type of copy number data provided. Currently supported options are a custom allele specific data format, and standard copy number VCF files.}

\item{maxgap}{The maximum gap between two breakpoints for them to be considered concordant. If unspecified, it is automatically set to 5 times the average interprobe distance of the assay.}

\item{save}{TRUE if want to export the shared mutations.}
}
\value{
List of shared breakpoints per pair It will generate a tsv file per pair in outdir with shared breakpoints.
}
\description{
Generate the shared breakpoints per pair calculated by breakclone and used to assess relatedness.
}
