% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{getSharedMuts}
\alias{getSharedMuts}
\title{Get common mutations}
\usage{
getSharedMuts(
  mutationTable,
  pairs,
  outdir = ".",
  scaleAFs = FALSE,
  save = FALSE
)
}
\arguments{
\item{mutationTable}{A segment table generated by the readAlleleSpecific or readVCFCn functions.}

\item{pairs}{A table of paired samples from the dataset. All tumours present in this table will be paired with all tumours from other patients.}

\item{outdir}{A path to save the shared mutations If unspecified, it is automatically set to the working directory.}

\item{scaleAFs}{Scale AFs per-sample by the highest AF within each sample. Only recommended for data with significant normal contamination that you are confident contains at least one clonal mutation per sample.}

\item{save}{TRUE if want to export the shared mutations.}
}
\value{
List of shared mutations per pair It will generate a tsv file per pair in outdir with shared mutations.
}
\description{
Get common mutations
}
