% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber.R
\name{makeReferenceCN}
\alias{makeReferenceCN}
\title{Generate reference distribution from copy number data}
\usage{
makeReferenceCN(
  segmentTable,
  pairs,
  patients = NULL,
  delimiter = NULL,
  cnType = c("alleleSpecific", "VCF"),
  excludeChromosomes = "Y",
  maxgap = NULL
)
}
\arguments{
\item{segmentTable}{A segment table generated by the readAlleleSpecific or readVCFCn functions.}

\item{pairs}{A table of paired samples from the dataset. All tumours present in this table will be paired with all tumours from other patients.}

\item{patients}{A character vector of patient IDs, parallel to the pairs table, used to prevent tumours originating from the same patient from being used in the reference distribution (optional).}

\item{delimiter}{A character separating patient IDs from tumour-specific identifiers in the sample IDs. Ignored if \code{patients} is provided (optional).}

\item{cnType}{The type of copy number data provided. Currently supported options are a custom allele specific data format, and standard copy number VCF files.}

\item{excludeChromosomes}{The name(s) of any chromosomes to be excluded.}

\item{maxgap}{The maximum gap between two breakpoints for them to be considered concordant. If unspecified, it is automatically set to 5 times the average interprobe distance of the assay.}
}
\value{
A numeric vector of pair scores comprising the reference distribution.
}
\description{
Generates the reference distribution of concordance scores from unpaired tumours for a given dataset.
}
\details{
The default is to assume one tumour pair per patient. If the sample is found in more than one pair, indicating more than one pair per patient, an error will be thrown.
In that case you must specify either \code{patients} or \code{delimiter}.
}
