% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{makeReferenceMutations}
\alias{makeReferenceMutations}
\title{Generate reference distribution from mutation data}
\usage{
makeReferenceMutations(
  mutationTable,
  pairs,
  patients = NULL,
  delimiter = NULL,
  additionalMutations = NULL,
  nAdditionalSamples = 0,
  excludeChromosomes = "Y",
  scaleAFs = FALSE
)
}
\arguments{
\item{mutationTable}{A table of mutations in each sample and their allele frequencies.}

\item{pairs}{A table of paired samples from the dataset. All tumours present in this table will be paired with all tumours from other patients.}

\item{patients}{A character vector of patient IDs, parallel to the pairs table, used to prevent tumours originating from the same patient from being used in the reference distribution (optional)}

\item{delimiter}{A character separating patient IDs from tumour-specific identifiers in the sample IDs. Ignored if \code{patients} is provided.}

\item{additionalMutations}{A table of mutations to be taken into account when calculating population frequencies. At a minimum, a table of the mutations in the population being studied. This is more informative when tumour type-specific mutations are included from external sources (e.g. TCGA).}

\item{nAdditionalSamples}{The number of samples used to derive the additional mutations table.}

\item{excludeChromosomes}{The name(s) of any chromosomes to be excluded.}

\item{scaleAFs}{Scale AFs per-sample by the highest AF within each sample. Only recommended for data with significant normal contamination that you are confident contains at least one clonal mutation per sample.}
}
\value{
A numeric vector of pair scores comprising the reference distribution.
}
\description{
Generates the reference distribution of concordance scores from unpaired tumours for a given dataset.
}
