% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutationTable}
\alias{mutationTable}
\title{mutationTable}
\format{
A data table with 5 columns:
\describe{
  \item{SampleID}{sample ID column}
  \item{Chr}{chromosome name column}
  \item{Pos}{coordinate column}
  \item{AF}{allele frequency column}
  \item{annotation}{mutation details column}
}
}
\usage{
mutationTable
}
\description{
List of mutation in a set of DCIS primaries and recurrences tumours, produced by sequencing.
}
\keyword{datasets}
