% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCNpairVCF}
\alias{plotCNpairVCF}
\title{Plot a VCF copy number pair with breakpoints.}
\usage{
plotCNpairVCF(
  binnedTable,
  cnTable,
  pair,
  segmentTable,
  breaks = NULL,
  colors = c("#f1562f", "#8a4e97"),
  limits = c(-1.5, 2),
  build = c("hg38", "hg19"),
  excludeChromosomes = "Y",
  fontLabelSize = 7,
  maxgap = NULL,
  sharedBarSize = 30
)
}
\arguments{
\item{binnedTable}{A segmented copy number matrix with bins per row and samples per column.}

\item{cnTable}{A raw copy number matrix with bins per row and samples per column.}

\item{pair}{A character vector with sample1 and sample2 IDs.}

\item{segmentTable}{A segment table generated by the breakclone::readAlleleSpecific or breakclone::readVCFCn functions. Also a list with a segment table of each type to plot both kind of data together.}

\item{breaks}{List of shared breakpoints per pair It will generate a tsv file per pair in outdir with shared breakpoints.}

\item{colors}{A character vector with colors for sample 1 and sample 2.}

\item{limits}{A numeric vector with y axis limits.}

\item{build}{The genome build of your copy number data. hg19 and hg38 work, other builds haven't been tested.}

\item{excludeChromosomes}{The name(s) of any chromosomes to be excluded.}

\item{fontLabelSize}{Number indicating size of the labels.}

\item{maxgap}{The maximum gap between two breakpoints for them to be considered concordant. If unspecified, it is automatically set to 5 times the average interprobe distance of the assay.}

\item{sharedBarSize}{Width of the shared breakpoints bars.}
}
\value{
Copy number plot.
}
\description{
Plot a VCF copy number pair with breakpoints.
}
