% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPie}
\alias{plotPie}
\title{Plot pie chart}
\usage{
plotPie(summary, colorClonality = c("#F5C61AD9", "#FB4F14D9", "#660C21D9"))
}
\arguments{
\item{summary}{A data frame listing the clonality results \code{clonalityResults}, clonality verdict based on the thresholds and number and fraction of shared and private breakpoints per sample.}

\item{colorClonality}{A character vector with colors for related, ambiguous and unrelated samples.}
}
\value{
Pie chart
}
\description{
Plot pie chart
}
\author{
Maria Roman Escorza
\email{maria.roman-escorza@kcl.ac.uk}
}
