% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotScatterVAF}
\alias{plotScatterVAF}
\title{VAF scatter plot for mutations in a pair.}
\usage{
plotScatterVAF(
  mutationTable,
  pair,
  title = "",
  xlab = "Primary",
  ylab = "Recurrence",
  colors = setNames(c("#3987bb", "#aee7ea", "#0c0d0c"), c(paste0(xlab, " private"),
    paste0(ylab, " private"), "Shared")),
  fontLabelSize = 7,
  limits = c(0, 1),
  dotSize = 2,
  annotGenes = FALSE,
  fontAnnotSize = 3,
  scaleAFs = FALSE
)
}
\arguments{
\item{mutationTable}{A table of mutations in each sample and their allele frequencies.}

\item{pair}{A segment table generated by the breakclone::readAlleleSpecific or breakclone::readVCFCn functions. Also a list with a segment table of each type to plot both kind of data together.}

\item{title}{A character vector with the directory to the shared breakpoints exported with exportSharedBreaks() function.}

\item{xlab}{A numeric character vector with the bin size.}

\item{ylab}{The type of copy number data provided. Currently supported options are a custom allele specific data format, and standard copy number VCF files.}

\item{colors}{A named color vector for private in sample 1, in sample 2 and shared breakpoints/mutations.}

\item{fontLabelSize}{Number indicating size of the labels.}

\item{dotSize}{Size of the dots.}

\item{annotGenes}{FALSE by default. TRUE if you want to make dot annotation. Extra column needs to be added to mutationTable if this is TRUE.}

\item{fontAnnotSize}{#' @param fontLabelSize Number indicating size of the annotated text.}

\item{scaleAFs}{Scale AFs per-sample by the highest AF within each sample. Only recommended for data with significant normal contamination that you are confident contains at least one clonal mutation per sample.}
}
\value{
Mutation plot.
}
\description{
VAF scatter plot for mutations in a pair.
}
