% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotScoresDensity}
\alias{plotScoresDensity}
\title{Plot the reference and pair distributions.}
\usage{
plotScoresDensity(reference, results)
}
\arguments{
\item{reference}{A reference distribution generated by one of the reference functions.}

\item{results}{A results table generated by one of the relatedness functions.}
}
\description{
Plot the reference and pair distributions.
}
