% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{readAlleleSpecific}
\alias{readAlleleSpecific}
\title{Read allele-specific copy number data from a directory}
\usage{
readAlleleSpecific(
  directory,
  pattern = "*_Segments_AbsCN_alleleSpecific.txt",
  sample.field = "SampleID",
  chr.field = "Chr",
  start.field = "Start",
  end.field = "End",
  nprobes.field = "nProbes",
  nmajor.field = "nMajor",
  nminor.field = "nMinor",
  ntotal.field = NULL
)
}
\arguments{
\item{directory}{A directory containing the files.}

\item{pattern}{The filename pattern matching the files.}

\item{sample.field}{Name of the sample ID column. If set to NULL, filenames will be used.}

\item{chr.field}{Name of the chromosome column.}

\item{start.field}{Name of the start coordinate column.}

\item{end.field}{Name of the end coordinate column.}

\item{nprobes.field}{Name of the per-segment number of probes column.}

\item{nmajor.field}{Name of the major copy number column.}

\item{nminor.field}{Name of the minor copy number column.}

\item{ntotal.field}{Name of the total copy number column (ignored if \code{nmajor.field} is set).}
}
\value{
A data.table with column names SampleID, Chr, Start, End, nProbes, nMajor, nMinor.
}
\description{
Read in allele-specific copy number data from tab- or comma- separated files. Required fields in each file are the chromosome name, start and end coordinates,
the number of probes in each segment, the minor copy number and either the major or total copy number.
}
