% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{readVCFCn}
\alias{readVCFCn}
\title{Read in copy number VCF files}
\usage{
readVCFCn(directory, pattern = "*.vcf")
}
\arguments{
\item{directory}{A directory containing the files.}

\item{pattern}{The filename pattern matching the files.}
}
\value{
A data.table with column names SampleID, Chr, Start, End, Bins, SVType, Length.
}
\description{
Read in copy number data from standard VCF files
}
