% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{readVCFMutations}
\alias{readVCFMutations}
\title{Read in mutation VCF files}
\usage{
readVCFMutations(directory, pattern = "*.vcf")
}
\arguments{
\item{directory}{A directory containing the files.}

\item{pattern}{The filename pattern matching the files.}
}
\value{
A data.table with column names SampleID, Chr, Pos, AF.
}
\description{
Read in mutation from standard VCF files
}
