% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segmentTable}
\alias{segmentTable}
\title{segmentTable}
\format{
A data table with 7 columns:
\describe{
  \item{SampleID}{sample ID column}
  \item{Chr}{chromosome name column}
  \item{Start}{start coordinate column}
  \item{End}{end coordinate column}
  \item{Bins}{per-segment number of probes column}
  \item{SVType}{type of CNA column. Either DUP for region of elevated copy number relative to the reference or DEL for deletion relative to the reference}
  \item{Length}{lenght of the segment column}
}
}
\usage{
segmentTable
}
\description{
Copy number aberrations in a set of DCIS primaries and recurrences tumours, produced by low-coverage Whole Genome Sequencing.
}
\keyword{datasets}
