% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segmentTable_AS}
\alias{segmentTable_AS}
\title{segmentTable_AS}
\format{
A data table with 7 columns:
\describe{
  \item{SampleID}{sample ID column}
  \item{Chr}{chromosome name column}
  \item{Start}{start coordinate column}
  \item{End}{end coordinate column}
  \item{nProbes}{per-segment number of probes column}
  \item{nMajor}{major copy number }
  \item{nMinor}{minor copy number }
}
}
\usage{
segmentTable_AS
}
\description{
Allele-specific copy number aberrations in a set of DCIS primaries and recurrences tumours, produced by SNParray.
}
\keyword{datasets}
