% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber.R
\name{summarizeClonalityCN}
\alias{summarizeClonalityCN}
\title{Summarize clonality results of copy number aberration data.}
\usage{
summarizeClonalityCN(
  clonalityResults,
  segmentTable,
  thres_ambiguous = 0.05,
  thres_related = 0.01,
  cnType = c("alleleSpecific", "VCF"),
  maxgap = NULL
)
}
\arguments{
\item{clonalityResults}{A data frame listing the tumour pairs contained in \code{pairs}, their relatedness scores and p-values for relatedness.}

\item{segmentTable}{A segment table generated by the readAlleleSpecific or readVCFCn functions.}

\item{thres_ambiguous}{P-value to define ambiguous. thres_ambiguous=0.05 by default.}

\item{thres_related}{P-value to define related thres_related=0.05 by default.}

\item{cnType}{The type of copy number data provided. Currently supported options are a custom allele specific data format, and standard copy number VCF files.}

\item{maxgap}{The maximum gap between two breakpoints for them to be considered concordant. If unspecified, it is automatically set to 5 times the average interprobe distance of the assay.}
}
\value{
A data frame listing the clonality results \code{clonalityResults}, clonality verdict based on the thresholds and number and fraction of shared and private breakpoints per sample.
}
\description{
Summarize clonality results of copy number aberration data.
}
