% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{summarizeClonalityMuts}
\alias{summarizeClonalityMuts}
\title{Summarize clonality results of mutation data.}
\usage{
summarizeClonalityMuts(
  clonalityResults,
  mutationTable,
  thres_ambiguous = 0.05,
  thres_related = 0.01,
  scaleAFs = FALSE
)
}
\arguments{
\item{clonalityResults}{A data frame listing the tumour pairs contained in \code{pairs}, their relatedness scores and p-values for relatedness.}

\item{mutationTable}{A table of mutations in each sample and their allele frequencies.}

\item{thres_ambiguous}{P-value to define ambiguous. thres_ambiguous=0.05 by default.}

\item{thres_related}{P-value to define related thres_related=0.05 by default.}

\item{scaleAFs}{Scale AFs per-sample by the highest AF within each sample. Only recommended for data with significant normal contamination that you are confident contains at least one clonal mutation per sample.}
}
\value{
A data frame listing the clonality results \code{clonalityResults}, clonality verdict based on the thresholds and number and fraction of shared and private mutations per sample.
}
\description{
Summarize clonality results of mutation data.
}
